+++ vulture_whitelist.py
@@ -1,0 +1,3 @@
+from moops import Group
+
+Group.custom  # noqa: B018